Updated graphics buffer handling.

+++ Fighting Game Engine/RenderingDataBuffer.cpp
@@ -6,8 +6,8 @@ void RenderingDataBuffer::SetupData(const glm::ivec2& screenSize, const glm::vec
 {
 	_data.resize(4, 0.0f);
 
-	_data[0] = screenSize.x;
-	_data[1] = screenSize.y;
+	_data[0] = (float)screenSize.x;
+	_data[1] = (float)screenSize.y;
 	_data[2] = screenSizeInverted.x;
 	_data[3] = screenSizeInverted.y;
 }
